third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/move-node-expected.txt

Tests elements drag and drop operation internals, verifies post-move selection.


Running: testDumpInitial
========= Original ========
- <div id="container">
      <div id="child1"></div>
      <div id="child2"></div>
      <div id="child3"></div>
      <div id="child4"></div>
  </div>

Running: testDragAndDrop
===== Moved child2 =====
- <div id="container">
      <div id="child1"></div>
      <div id="child3"></div>
      <div id="child2"></div>
      <div id="child4"></div>
  </div>
Selection: div#child2
